[Bugfix] cast to float32 before isnan on the fp8 weight_scale sentinel check
torch.isnan() on a raw float8_e4m3fn tensor works on CUDA (verified on
GB10), but isn't guaranteed portable across every backend vLLM supports
(ROCm/XPU/CPU fp8 kernels aren't required to implement isnan directly).
Casting to float32 first is a no-op in outcome (a NaN bit pattern stays
NaN after upcast) and removes the backend dependency.

Reported by CodeRabbit on #52501.

+++ vllm/model_executor/layers/quantization/modelopt.py
@@ -1968,7 +1968,7 @@ class KNvfp4Static(QuantKeyScheme):
         # means the FP4 weights were never actually loaded (e.g. the
         # checkpoint stores this layer as BF16 and the weight loader silently
         # skipped it).
-        if torch.isnan(layer.weight_scale).any():
+        if torch.isnan(layer.weight_scale.float()).any():
             raise RuntimeError(
                 f"NVFP4 weight_scale for layer "
                 f"{getattr(layer, 'name', repr(layer))!r} was never loaded "
